Added sync of last modified date from server and retry counter for deletions
The client sets the last modified date of a downloaded file/folder to the date
of the server side.

If a file is deleted locally, but could not be deleted on the server, it is skipped if
the remote operation failed 3 times. Also a local change on a write protected
file will be skipped after the third failed remote operation.